app/src/main/java/org/openlmis/core/network/model/RnrFormStatusEntry.java
/*
 * This program is part of the OpenLMIS logistics management information
 * system platform software.
 *
 * Copyright © 2015 ThoughtWorks, Inc.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published
 * by the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version. This program is distributed in the
 * hope that it will be useful, but WITHOUT ANY WARRANTY; without even the
 * implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the GNU Affero General Public License for more details. You should
 * have received a copy of the GNU Affero General Public License along with
 * this program. If not, see http://www.gnu.org/licenses. For additional
 * information contact info@OpenLMIS.org
 */

package org.openlmis.core.network.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.openlmis.core.model.RnRForm;
import org.openlmis.core.model.RnRForm.Status;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class RnrFormStatusEntry {

  private long id;
  private RnRForm.Status status;

  public boolean isValidStatus() {
    return status == Status.IN_APPROVAL
        || status == Status.APPROVED
        || status == Status.REJECTED;
  }
}